Add tests for new metric system
